Refactoring rpi code
Refactoring of the rpi code
